+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Scripts.hs
@@ -47,8 +47,9 @@ module Cardano.Ledger.Alonzo.Scripts (
   pattern SpendingPurpose,
   pattern MintingPurpose,
   pattern CertifyingPurpose,
+  pattern WithdrawingPurpose,
   pattern RewardingPurpose,
-  AlonzoPlutusPurpose (..),
+  AlonzoPlutusPurpose (.., AlonzoRewarding),
   AsPurpose (..),
   AsItem (..),
   AsIx (..),
@@ -200,14 +201,24 @@ class
 
   toCertifyingPurpose :: PlutusPurpose f era -> Maybe (f Word32 (TxCert era))
 
+  mkWithdrawingPurpose :: f Word32 AccountAddress -> PlutusPurpose f era
+
+  toWithdrawingPurpose :: PlutusPurpose f era -> Maybe (f Word32 AccountAddress)
+
   mkRewardingPurpose :: f Word32 AccountAddress -> PlutusPurpose f era
+  mkRewardingPurpose = mkWithdrawingPurpose
 
   toRewardingPurpose :: PlutusPurpose f era -> Maybe (f Word32 AccountAddress)
+  toRewardingPurpose = toWithdrawingPurpose
 
   upgradePlutusPurposeAsIx ::
     AlonzoEraScript (PreviousEra era) =>
     PlutusPurpose AsIx (PreviousEra era) ->
     PlutusPurpose AsIx era
+
+{-# DEPRECATED mkRewardingPurpose "In favor of `mkWithdrawingPurpose`" #-}
+
+{-# DEPRECATED toRewardingPurpose "In favor of `toWithdrawingPurpose`" #-}
 
 mkBinaryPlutusScript ::
   (MonadFail m, AlonzoEraScript era) =>
@@ -309,7 +320,7 @@ data AlonzoPlutusPurpose f era
   = AlonzoSpending !(f Word32 TxIn)
   | AlonzoMinting !(f Word32 PolicyID)
   | AlonzoCertifying !(f Word32 (TxCert era))
-  | AlonzoRewarding !(f Word32 AccountAddress)
+  | AlonzoWithdrawing !(f Word32 AccountAddress)
   deriving (Generic)
 
 deriving instance Eq (AlonzoPlutusPurpose AsIx era)
@@ -340,7 +351,7 @@ instance
     AlonzoSpending x -> rnf x
     AlonzoMinting x -> rnf x
     AlonzoCertifying x -> rnf x
-    AlonzoRewarding x -> rnf x
+    AlonzoWithdrawing x -> rnf x
 
 instance
   ( forall a b. (EncCBOR a, EncCBOR b) => EncCBOR (f a b)
@@ -354,7 +365,7 @@ instance
     AlonzoSpending p -> encodeWord8 0 <> encCBOR p
     AlonzoMinting p -> encodeWord8 1 <> encCBOR p
     AlonzoCertifying p -> encodeWord8 2 <> encCBOR p
-    AlonzoRewarding p -> encodeWord8 3 <> encCBOR p
+    AlonzoWithdrawing p -> encodeWord8 3 <> encCBOR p
 
 instance
   ( forall a b. (DecCBOR a, DecCBOR b) => DecCBOR (f a b)
@@ -369,7 +380,7 @@ instance
       0 -> AlonzoSpending <$> decCBOR
       1 -> AlonzoMinting <$> decCBOR
       2 -> AlonzoCertifying <$> decCBOR
-      3 -> AlonzoRewarding <$> decCBOR
+      3 -> AlonzoWithdrawing <$> decCBOR
       n -> fail $ "Unexpected tag for AlonzoPlutusPurpose: " <> show n
 
 deriving via
@@ -404,9 +415,13 @@ instance
     AlonzoSpending n -> kindObjectWithValue "AlonzoSpending" n
     AlonzoMinting n -> kindObjectWithValue "AlonzoMinting" n
     AlonzoCertifying n -> kindObjectWithValue "AlonzoCertifying" n
-    AlonzoRewarding n -> kindObjectWithValue "AlonzoRewarding" n
+    AlonzoWithdrawing n -> kindObjectWithValue "AlonzoWithdrawing" n
     where
       kindObjectWithValue name n = kindObject name ["value" .= n]
+
+pattern AlonzoRewarding :: f Word32 AccountAddress -> AlonzoPlutusPurpose f era
+pattern AlonzoRewarding x = AlonzoWithdrawing x
+{-# DEPRECATED AlonzoRewarding "In favor of `AlonzoWithdrawing`" #-}
 
 pattern SpendingPurpose ::
   AlonzoEraScript era => f Word32 TxIn -> PlutusPurpose f era
@@ -426,11 +441,18 @@ pattern CertifyingPurpose c <- (toCertifyingPurpose -> Just c)
   where
     CertifyingPurpose c = mkCertifyingPurpose c
 
+pattern WithdrawingPurpose ::
+  AlonzoEraScript era => f Word32 AccountAddress -> PlutusPurpose f era
+pattern WithdrawingPurpose c <- (toWithdrawingPurpose -> Just c)
+  where
+    WithdrawingPurpose c = mkWithdrawingPurpose c
+
 pattern RewardingPurpose ::
   AlonzoEraScript era => f Word32 AccountAddress -> PlutusPurpose f era
-pattern RewardingPurpose c <- (toRewardingPurpose -> Just c)
+pattern RewardingPurpose c <- (toWithdrawingPurpose -> Just c)
   where
-    RewardingPurpose c = mkRewardingPurpose c
+    RewardingPurpose c = mkWithdrawingPurpose c
+{-# DEPRECATED RewardingPurpose "In favor of `WithdrawingPurpose`" #-}
 
 -- Alonzo Script ===============================================================
 
@@ -546,7 +568,7 @@ instance AlonzoEraScript AlonzoEra where
     AlonzoSpending x -> AlonzoSpending $ f x
     AlonzoMinting x -> AlonzoMinting $ f x
     AlonzoCertifying x -> AlonzoCertifying $ f x
-    AlonzoRewarding x -> AlonzoRewarding $ f x
+    AlonzoWithdrawing x -> AlonzoWithdrawing $ f x
 
   mkSpendingPurpose = AlonzoSpending
 
@@ -563,10 +585,10 @@ instance AlonzoEraScript AlonzoEra where
   toCertifyingPurpose (AlonzoCertifying i) = Just i
   toCertifyingPurpose _ = Nothing
 
-  mkRewardingPurpose = AlonzoRewarding
+  mkWithdrawingPurpose = AlonzoWithdrawing
 
-  toRewardingPurpose (AlonzoRewarding i) = Just i
-  toRewardingPurpose _ = Nothing
+  toWithdrawingPurpose (AlonzoWithdrawing i) = Just i
+  toWithdrawingPurpose _ = Nothing
 
   upgradePlutusPurposeAsIx =
     error "Impossible: No `PlutusScript` and `AlonzoEraScript` instances in the previous era"